Manual Fix

+++ RogueEngine/Source/PlayerControllerSystem.cpp
@@ -98,7 +98,7 @@ namespace Rogue
 					{
 						//rigidbody.addForce(Vec2(100.0f, 0.0f) * g_deltaTime * 1000.0f);
 						//RE_INFO("Move A Right!");
-						Event* ev = new EntMoveEvent{ (int)*iEntity, true, 100.0f, 0.0f };
+						Event* ev = new EntMoveEvent{ *iEntity, true, 100.0f, 0.0f };
 						ev->SetSystemReceivers((int)SystemID::id_PHYSICSSYSTEM);
 						EventDispatcher::instance().AddEvent(ev);
 					}
